Silence MSVS warnings.

+++ Source/Parameter.cpp
@@ -147,7 +147,7 @@ void ParamStepListenFreq::changeListenerCallback (ChangeBroadcaster* source)
             // calc temp values
         const float numStepsPower = numStepsParam.get();
         const float newNumSteps = 1.5f * std::pow (2.0f, numStepsPower); // for [1:5] yields [3:48]
-        setNumSteps (newNumSteps);
+        setNumSteps (static_cast<int> (newNumSteps));
 
             // set value for parameter (setValueNotifyingHost expects 0to1)
         const float val0to1 = getRange().convertTo0to1 (get());
